hexen: Fix clang abs warning in R_StoreWallRange

+++ src/hexen/r_segs.c
@@ -348,7 +348,7 @@ void R_StoreWallRange(int start, int stop)
 // calculate rw_distance for scale calculation
 //
     rw_normalangle = curline->angle + ANG90;
-    offsetangle = abs(rw_normalangle - rw_angle1);
+    offsetangle = abs((int) rw_normalangle - (int) rw_angle1);
     if (offsetangle > ANG90)
         offsetangle = ANG90;
     distangle = ANG90 - offsetangle;
